first_hit: return int8 sides again
The result is only -1 / 0 / 1 and P&L is the backtester's job, so the
compact int8 array is the right type; the docstring says to cast with
.astype(np.int64) before doing arithmetic on it.

+++ tests/test_tick_data.py
@@ -139,7 +139,7 @@ def test_arrays_match_single_calls(data, freq):
     hi = p0 + rng.integers(0, 80, len(starts))
     lo = p0 - rng.integers(0, 80, len(starts))
     sides = first_hit(data, freq, starts, hi, lo)
-    assert sides.dtype == np.int64  # safe for P&L arithmetic (sides * tp)
+    assert sides.dtype == np.int8
     assert sides.tolist() == [first_hit(data, freq, int(s), int(h), int(l)) for s, h, l in zip(starts, hi, lo)]
